+++ 6583XA-Proto-CutDown/Src/mc_tasks_foc.c
@@ -606,8 +606,8 @@ inline uint16_t FOC_CurrControllerM1(void)
   RCM_ExecNextConv();
   Ialphabeta = MCM_Clarke(Iab);
   Iqd = MCM_Park(Ialphabeta, hElAngle);
-  Vqd.q = PI_Controller(pPIDIq[M1], (int32_t)(FOCVars[M1].Iqdref.q) - Iqd.q);
-  Vqd.d = PI_Controller(pPIDId[M1], (int32_t)(FOCVars[M1].Iqdref.d) - Iqd.d);
+  Vqd.q = PI_Controller(pPIDIq[M1], (FOCVars[M1].Iqdref.q) - Iqd.q);
+  Vqd.d = PI_Controller(pPIDId[M1], (FOCVars[M1].Iqdref.d) - Iqd.d);
   Vqd = Circle_Limitation(&CircleLimitationM1, Vqd);
   hElAngle += SPD_GetInstElSpeedDpp(speedHandle)*REV_PARK_ANGLE_COMPENSATION_FACTOR;
   Valphabeta = MCM_Rev_Park(Vqd, hElAngle);
